feat: create data transfer objects for user service methods
Create `dto` folder and create inside it three files:

- `create-user.input.ts` to define `SignUpUserInput` class that specifies
the type of the data that should come  from the client side to create
or sign up users.

- `update-data-user.input.ts` to define `UpdateUserDataInput` class that
specifies the type of the data that should come from the client side to
update users data.

- `search-for-users.input.ts` to define `SearchForUserInput` class that
specifies how the client side can search for users.