test: create tests folders & configuration files to organize the tests
Rename `test` folder to `tests` and create inside this folder three other folders:

- `e2e` folder for collecting the "end to end" tests inside it, Also create inside
it `jest-e2e.json` file which is the configuration file of "end to end" tests.

- `integration` folder for collecting integration tests inside it, Also create inside
it `jest-e2e.json` file which is the configuration file of integration tests.

- `units` folder for collecting units tests inside it, Also create inside
it `jest-e2e.json` file which is the configuration file of units tests.