test: create the functions that prepare testing modules
Create `createTestingApp.ts` function in `tests` folder to provide
a shorter way to create a testing instance of the app module for
using it in "End-to-end" tests.

Create `createTestingService.ts` function in `tests` folder to provide
a shorter way to create a testing instance of a service class (for
example `UsersService`) for using it in "Integration" tests.

Create `testingDatabaseURI.ts` file in `tests` folder to provide
the shared URI of the testing database (mongo).