feat: change all user's `avatar` fields types to the new shape of the field
- Create `UserAvatar` object type in `src/users/dto/user-data.type.ts` file
and convert the type of `avatar` field of `User` object type in this file to
`UserAvatar`.

- Use `UserAvatar` from `src/users/dto/user-data.type.ts` file as the type
of `avatar` field of `CommentOwner` object type in
`src/post-comments/dto/comment.type.ts` file.

- Create `UpdateAvatarInput` object type and set it as type of `avatar` filed
of `UpdateUserDataInput` class in `src/users/dto/update-data-user.input.ts`
file.

- Create `src/users/schemas/user-avatar.schema.ts` file and define inside
it `AvatarSchema` schema type and import it to use it as type of `avatar`
field of `User` schema in `src/users/schemas/user.schema.ts` file.